Deal with 2 different kinds of magic SVs in tc_Int.
This commit fix tests related to validate Int in tied array/scalar,
on perl5.8 ... perl5.16.

The 2 types of magic SVs are:

1. SvTYPE(sv) == SVt_PVMG
2. SvMAGIC(sv)

It appears to me that, when they are assigned integer values, -- even
if it's originated from integer literals -- there are no "public
integer" inside these magical SV, only "private" ones.

We have been using SvIOKp(sv) to probe the Integer-ness of an SV and
that just happen to be seemingly the only ways to do so for magic SVs.